# apps/backend/internal/docs/swagger.yaml
basePath: /api/v1
definitions:
  api_http_handlers_alert.CreateAlertWithChannelsParams:
    properties:
      ca:
        example: SP1Y5YSTAHZ88XYK1VPDH24GY0HPX5J4JECTMY4A1.velar-token
        type: string
      channels:
        items:
          type: string
        type: array
      metric:
        type: string
      operator:
        type: string
      repeatable:
        type: boolean
      userId:
        type: string
      value:
        type: string
    required:
    - ca
    - metric
    - operator
    - repeatable
    - value
    type: object
  backend_api_http.InternalServerErrorResponse:
    properties:
      errors: {}
      message:
        type: string
      status:
        example: Internal Server Error
        type: string
    type: object
  backend_api_http.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status:
        type: string
    type: object
  backend_internal_db_repository.Alert:
    properties:
      ca:
        example: SP1Y5YSTAHZ88XYK1VPDH24GY0HPX5J4JECTMY4A1.velar-token
        type: string
      createdAt:
        type: string
      id:
        type: string
      metric:
        type: string
      operator:
        type: string
      repeatable:
        type: boolean
      status:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
      value:
        type: string
    required:
    - ca
    - metric
    - operator
    - repeatable
    - status
    - value
    type: object
  backend_internal_db_repository.Channel:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  backend_internal_db_repository.UpdateAlertWithChannelsParams:
    properties:
      alertId:
        type: string
      channelIds:
        items:
          type: string
        type: array
      metric:
        type: string
      operator:
        type: string
      repeatable:
        type: boolean
      userId:
        type: string
      value:
        type: string
    type: object
  backend_internal_db_repository.UpdateAlertWithChannelsRow:
    properties:
      ca:
        type: string
      channels:
        items:
          type: integer
        type: array
      createdAt:
        type: string
      id:
        type: string
      metric:
        type: string
      operator:
        type: string
      repeatable:
        type: boolean
      status:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
      value:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: atomic.k.2739@gmail.com
    name: Atomic
    url: https://github.com/iatomic1
  description: API for Dexion
  title: DEXION API
  version: "1.0"
paths:
  /alerts:
    get:
      consumes:
      - application/json
      description: Fetch all alerts belonging to the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: User alerts retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/backend_api_http.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/backend_internal_db_repository.Alert'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_api_http.InternalServerErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Retrieve all alerts for the authenticated user
      tags:
      - Alerts
    post:
      consumes:
      - application/json
      description: Create a new alert with optional notification channels
      parameters:
      - description: Alert data with channels
        in: body
        name: AlertRequest
        required: true
        schema:
          $ref: '#/definitions/api_http_handlers_alert.CreateAlertWithChannelsParams'
      produces:
      - application/json
      responses:
        "201":
          description: Alert created successfully
          schema:
            allOf:
            - $ref: '#/definitions/backend_api_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/backend_internal_db_repository.Alert'
              type: object
        "400":
          description: Invalid request data
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Alert already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_api_http.InternalServerErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new alert
      tags:
      - Alerts
  /alerts/{id}:
    delete:
      consumes:
      - application/json
      description: Remove an alert belonging to the authenticated user
      parameters:
      - description: Alert ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Alert deleted successfully
          schema:
            allOf:
            - $ref: '#/definitions/backend_api_http.Response'
            - properties:
                data: {}
              type: object
        "400":
          description: Invalid UUID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Alert not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_api_http.InternalServerErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete an existing alert
      tags:
      - Alerts
    get:
      consumes:
      - application/json
      description: Fetch a single alert by its UUID for the authenticated user
      parameters:
      - description: Alert ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Alert retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/backend_api_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/backend_internal_db_repository.Alert'
              type: object
        "400":
          description: Invalid UUID format
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_api_http.InternalServerErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Retrieve a specific alert by ID
      tags:
      - Alerts
    patch:
      consumes:
      - application/json
      description: Update the details of an alert belonging to the authenticated user
      parameters:
      - description: Alert ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Alert update data
        in: body
        name: Alert
        required: true
        schema:
          $ref: '#/definitions/backend_internal_db_repository.UpdateAlertWithChannelsParams'
      produces:
      - application/json
      responses:
        "200":
          description: Alert updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/backend_api_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/backend_internal_db_repository.UpdateAlertWithChannelsRow'
              type: object
        "400":
          description: Invalid UUID format or request data
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Alert not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_api_http.InternalServerErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update an existing alert
      tags:
      - Alerts
  /alerts/channels:
    get:
      consumes:
      - application/json
      description: Fetch all available channels in the system
      produces:
      - application/json
      responses:
        "200":
          description: Channels retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/backend_api_http.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/backend_internal_db_repository.Channel'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_api_http.InternalServerErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Retrieve all channels
      tags:
      - Alerts
...
schemes:
- http
- https
securityDefinitions:
  AccessTokenBearer:
    description: AccessTokenBearer Authentication
    in: header
    name: Authorization
    type: apiKey
  Google OAuth2:
    authorizationUrl: https://accounts.google.com/o/oauth2/auth
    flow: accessCode
    scopes:
      email: Grants access to user's email
      profile: Grants access to user's basic profile info
    tokenUrl: https://oauth2.googleapis.com/token
    type: oauth2
  RefreshTokenBearer:
    description: RefreshTokenBearer Authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...